ayther (#26): la mascara de mute pasa a 4 bytes y el chip PCM entra en ella
`ayther_audio_mute` eran 16 bits: seis para el FM, cuatro para el PSG. Los ocho
canales del RF5C164 de Sega CD no entraban.

DEJARLO AFUERA NO ERA NEUTRAL. Este core ya tiene un SEGUNDO camino de silencio
que si soporta PCM —`audio_probe_set_channel_gain`, aplicado en el mixer de
pcm.c— asi que un frontend que quiera callar ese chip para sustituirlo se ve
empujado a usarlo. Y tener la decision de mute escrita en dos lugares es
exactamente el defecto que el frontend corrigio en su issue #329: la voz del
chip callaba por un camino y el reemplazo seguia sonando por el otro. Un solo
mecanismo, mas ancho.

  bits  0-5   FM  (YM2612)   0-5
  bits  6-9   PSG (SN76489)  0-3
  bits 10-17  PCM (RF5C164)  0-7     <- nuevo
  bits 18-31  libres, y el core los RECHAZA al escribir

El mute del PCM se aplica DESPUES de leer la muestra y avanzar `addr`, a
proposito: el chip tiene que evolucionar igual con el mute puesto y sin el, o el
replay divergiria. Sale a cero solo lo que se suma al mixer — el mismo criterio
que ya usaban el FM y el PSG.

La validacion de bits reservados de write_control se comparaba contra un
`sizeof(uint16_t)` escrito a mano. Con la region en 4 bytes esa rama nunca se
habria cumplido y el chequeo habria quedado MUERTO sin que nadie lo notara;
ahora compara contra el `byte_size` de la propia region.

MEDIDO desde el frontend (Earthworm Jim SE, 3600 frames, tres corridas del mismo
tramo):

  rms  A1=1852.0   A2=1852.0 (control A-vs-A)   B=328.4 (los 8 PCM muteados)

El PCM aportaba el 82,3% de la energia del tramo. El control A-vs-A no es
ceremonia: sin el, una corrida que difiriera por deriva del emulador pasaria por
"el mute funciona". Y B > 0 confirma que el FM y el PSG siguen sonando, o sea
que no se convirtio en un mute global.

Gotcha que costo una corrida en rojo: el mute esta gateado por
AYTHER_SUB_RENDER_CONTROLS. Un runner recien creado no esta suscripto a nada y
la mascara se ignora en silencio — como debe ser, pero conviene saberlo.

Suite del fork: 83 + 10 + 23 + 18 + 13 checks, verdes.

+++ core/sound/sound.c
@@ -44,7 +44,7 @@ int8 audio_hard_disable = 0;
 
 #ifdef AYTHER_EXTENSIONS
 /* AYTHER fork delta: máscara de mute por canal (ver sound.h). 0 = todo suena. */
-uint16 ayther_audio_mute = 0;
+uint32 ayther_audio_mute = 0;
 #endif
 
 /* YM2612 internal clock = input clock / 6 = (master clock / 7) / 6 */
